use a smaller dc in case the test is failing due to resource constraints
The test is passing locally for me so it is entirely possibly that the failure
in GHA is due to resource constraints. There is no need to deploy 6 C* nodes
to test a config change particularly when the tests are running in GHA.
Furthermore, it will just make the test slower.